# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Social Network",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://social.loc/api"
        }
    ],
    "paths": {
...
        "/user": {
            "get": {
                "tags": [
                    "user"
                ],
                "summary": "Get auth user",
                "operationId": "App\\Http\\Controllers\\UserController::getAuthUser",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "example": {
                                                "id": 1,
                                                "name": "John",
                                                "surname": "Doe",
                                                "avatar": "http://site.com/img/avatar.png",
                                                "friends_count": 355,
                                                "friend_requests_count": 6,
                                                "email": "user@gmail.com",
                                                "created_at": "2021-01-15 12:59:20",
                                                "updated_at": "2021-01-20 04:25:15",
                                                "join_date": "10 days ago"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
        "/users/{user}": {
            "get": {
                "tags": [
                    "users"
                ],
                "summary": "Get user by id",
                "description": "Get user with relationship info",
                "operationId": "App\\Http\\Controllers\\UserController::show",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "User id to fetch",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "example": {
                                                "id": 1,
                                                "avatar": "http://site.com/img/default.png",
                                                "friends_count": 0,
                                                "name": "Test",
                                                "surname": "First",
                                                "relationship": {
                                                    "id": 8,
                                                    "status": "pending",
                                                    "is_sender": false
                                                },
                                                "created_at": "2021-01-16T13:39:29",
                                                "updated_at": "2021-01-16T13:39:29",
                                                "join_date": "4 days ago"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/user/friends?page={page}": {
            "get": {
                "tags": [
                    "user"
                ],
                "summary": "Get auth user friends",
                "operationId": "App\\Http\\Controllers\\UserController::getFriends",
                "parameters": [
                    {
                        "name": "page",
                        "in": "path",
                        "description": "Paginates data",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserCollection"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "UserCollection": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/User"
                        }
                    },
                    "links": {
                        "type": "object",
                        "example": {
                            "first": "first page link",
                            "last": "last page link",
                            "next": "next page link",
                            "prev": "prev page link"
                        }
                    },
                    "meta": {
                        "type": "object",
                        "example": {
                            "current_page": 1,
                            "from": 1,
                            "last_page": 1,
                            "path": "full url path",
                            "per_page": 10,
                            "to": 2,
                            "total": 2
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "UserCollection"
                }
            },
...
            "User": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "name": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "John"
                    },
                    "surname": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Doe"
                    },
                    "avatar": {
                        "type": "string",
                        "readOnly": "true",
                        "example": "http://site.com/img/avatar.png"
                    },
                    "friends_count": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "355"
                    },
                    "email": {
                        "description": "User unique email address",
                        "type": "string",
                        "format": "email",
                        "readOnly": "true",
                        "example": "user@gmail.com"
                    },
                    "created_at": {
                        "description": "Registration date and time",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true",
                        "example": "2021-01-15 12:59:20"
                    },
                    "updated_at": {
                        "description": "Last updated date and time",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true",
                        "example": "2021-01-20 04:25:15"
                    },
                    "join_date": {
                        "description": "Registration date in human radable format",
                        "type": "string",
                        "readOnly": "true",
                        "example": "10 days ago"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "User"
                }
            }
        },
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "description": "Use /login endpoint to get access token",
                "name": "Bearer",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}